Add primitiveType = this.nodeTypePrimitiveType[nodeType]
Used in eliminateNodes()

Also comment out "break;" in parse while loop to
replace all occurrences of same pattern before
trying all grammar rules all over again from the beginning.
This is to fix the JSON example, but it will break the
derived grammar for Rust, so we need to think about this.