Include most recently replaced node in the new nodeString to match
This allows replacing all occurrences in the same precedence group in one sweep,
even if the nodeType of the new node that replaces a pattern of nodes
have children with the same nodeType as the new node.

Example: Binary = (Lit Add Lit|Binary Add Lit)

Now both the JSON and Rust demo works.