Add fixPrecedenceGroups() which automatically derive precedence for grammar rules
The strategy is to detect the first invalid node added by parse().
This is implemented by passing expectedParseTree to parse(),
and then to check if each new node added by parse() is valid
by passing it to the new function parseTreeContainsSubParseTree().

The new function parseTreeContainsSubParseTree() takes the valid parse tree
as first argument, and the sub tree of the new node to check as second argument.
It walks the tree recursively and checks if any of the nodes in the valid parse tree
matches the sub tree exactly.

To prevent some other part of the valid parse tree to match where it structurally is identical,
but actually consists of different token node leaves, each token node leave had to be given
a unique tokenId.

grammarRules has been replaced by exportGrammarRules(),
which creates the same structure used to import grammar rules via loadGrammarRules()
by reading from tokenizerGrammarRules and parserGrammarRuleIdsByParserAndPrecedenceGroup.
This was necessary as fixPrecedenceGroups() makes modifications to this.parserGrammarRuleIdsByParserAndPrecedenceGroup,
that must remain when exporting.

Before, precedence '' (empty string) meant all grammar rules had different precedence.
This has now been changed to mean all the grammar rules have the same precedence,
which is a better starting position, as fixPrecedenceGroups() moves the grammar rule that matched too early
out from the precedence group.